requirements/base.txt
Django==1.7.5
argparse==1.2.1
djangorestframework==3.0.5
django-rest-swagger==0.2.8
wsgiref==0.1.2
python-magic==0.4.6
jose==0.2.2
pycrypto==2.6.1
pygpgme==0.3
django-import-export==0.2.6
djangorestframework-jwt==1.2.0
djoser==0.2.0
django-filter==0.9.2
djangorestframework-filters==0.2.2
django-grappelli==2.6.3
